+++ pom.xml
@@ -9,7 +9,7 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.example</groupId>
-	<artifactId>EmployeeManagementSystem</artifactId>
+	<artifactId>EmployeeMgmtSystem</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>spring</name>
 	<description>Demo project for Spring Boot</description>
