+++ src/main/java/com/example/EmployeeMgmtSystem/Employee.java
@@ -1,6 +1,6 @@
 /* Accessing data with MySQL */
 
-package com.example.EmployeeManagementSystem;
+package com.example.EmployeeMgmtSystem;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
