+++ src/main/java/com/example/EmployeeMgmtSystem/EmployeeRepository.java
@@ -1,6 +1,6 @@
 /* Accessing data with MySQL */
 
-package com.example.EmployeeManagementSystem;
+package com.example.EmployeeMgmtSystem;
 
 import org.springframework.data.repository.CrudRepository;
