+++ src/main/java/com/example/EmployeeMgmtSystem/MessageStatus.java
@@ -1,3 +1,3 @@
-package com.example.EmployeeManagementSystem;
+package com.example.EmployeeMgmtSystem;
 
 public record MessageStatus(String status_type, String status_msg) {}
