+++ src/test/java/com/example/EmployeeMgmtSystem/ApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.EmployeeManagementSystem;
+package com.example.EmployeeMgmtSystem;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
